requirements.txt
fastapi
uvicorn
sqlalchemy
websockets
requests
python-dotenv
google-genai
undetected-chromedriver
selenium
